refactor contents of FeedItem
