+++ components/ui/Feed/FeedItem/index.ts
@@ -1,0 +1,1 @@
+export { default as FeedItem } from "./FeedItem";
